+++ CHANGES.md
@@ -3,6 +3,8 @@
 Unreleased
 
 -   Drop support for Python 3.9.
+-   Add `get_or_abort` and `one_or_abort` methods, which get a single row or
+    otherwise tell Flask to abort with a 404 error.
 
 ## Version 0.1.0